Refactor WaitForNetworkIdle
